Add smooth trajectory for the ZMP reference and the swinging foot (#1)
* Add a new method to generate swing foot path

* Allow to change the feet trajectory generation method

* Small fixes
This PR provides two improvements:

The ZMP reference path is now generated using cubic spline.
The swing foot trajectory is generated using a minimal jerk Bezier curve.

This create a smoother motion of the robot and reduces the impact forces when the robot touches the floor. I maintained the old path generation method for comparison purpose. The user can specify a function to generate it.

Additionally this PR provides slight improvement of the CI.

The figures below show the difference the swing foot path generated before and after the dev: